# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Virtual Bank API",
      "version": "2.1"
    },
    "paths": {
...
      "/accounts": {
        "get": {
          "tags": [
            "2 Счета и балансы"
          ],
          "summary": "1. Получить список счетов",
          "description": "## 💳 Получение списка счетов клиента\n\n### Два режима работы:\n\n#### 1️⃣ Запрос своих счетов (в том же банке)\n```bash\nGET /accounts\nHeaders:\n  Authorization: Bearer \u003Cclient_token\u003E\n```\n\n#### 2️⃣ Межбанковый запрос (с согласием)\n```bash\nGET /accounts?client_id=cli-ab-001\nHeaders:\n  Authorization: Bearer \u003Cbank_token\u003E\n  X-Requesting-Bank: team200\n  X-Consent-Id: \u003Cconsent_id\u003E\n```\n\n### Ответ содержит:\n- `account_id` — уникальный идентификатор счета\n- `currency` — валюта (RUB, USD, EUR)\n- `account_type` — тип счета (Personal, Business)\n- `nickname` — название счета\n- `servicer` — информация о банке\n\n### ⚠️ Важно для межбанковых запросов:\n1. Сначала создайте согласие: `POST /account-consents/request`\n2. Клиент должен одобрить согласие в банке-владельце счетов\n3. Используйте полученный `consent_id` в заголовке `X-Consent-Id`\n4. Укажите свой банк в `X-Requesting-Bank`\n\n### Примечание:\n- Без согласия межбанковый запрос вернет 403 с подсказкой, как получить согласие\n- Согласие имеет срок действия (обычно 90 дней)",
          "operationId": "get_accounts_accounts_get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "client_id",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID клиента (например team200-1). Обязателен для межбанковых запросов",
                "title": "Client Id"
              },
              "description": "ID клиента (например team200-1). Обязателен для межбанковых запросов",
              "example": "team200-1"
            },
            {
              "name": "x-consent-id",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID согласия (получите через POST /account-consents/request). Обязателен для межбанковых запросов",
                "title": "X-Consent-Id"
              },
              "description": "ID согласия (получите через POST /account-consents/request). Обязателен для межбанковых запросов",
              "example": "consent-69e75facabba"
            },
            {
              "name": "x-requesting-bank",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID вашей команды (от организаторов). Укажите для запроса данных из другого банка",
                "title": "X-Requesting-Bank"
              },
              "description": "ID вашей команды (от организаторов). Укажите для запроса данных из другого банка",
              "example": "team200"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "2 Счета и балансы"
          ],
          "summary": "5. Создать счет",
          "description": "Создание нового счета\n\nПоддерживаемые типы: checking, savings",
          "operationId": "create_account_accounts_post",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/accounts/{account_id}": {
        "get": {
          "tags": [
            "2 Счета и балансы"
          ],
          "summary": "2. Получить детали счета",
          "description": "Получение детальной информации о счете\n\n**Требует:** Client token (для своих счетов) или Bank token с согласием (межбанк)",
          "operationId": "get_account_accounts__account_id__get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "account_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "description": "ID счета",
                "title": "Account Id"
              },
              "description": "ID счета",
              "example": "acc-1010"
            },
            {
              "name": "x-consent-id",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID согласия (получите через POST /account-consents/request). Обязателен для межбанковых запросов",
                "title": "X-Consent-Id"
              },
              "description": "ID согласия (получите через POST /account-consents/request). Обязателен для межбанковых запросов",
              "example": "consent-69e75facabba"
            },
            {
              "name": "x-requesting-bank",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID вашей команды (от организаторов). Укажите для запроса данных из другого банка",
                "title": "X-Requesting-Bank"
              },
              "description": "ID вашей команды (от организаторов). Укажите для запроса данных из другого банка",
              "example": "team200"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/accounts/{account_id}/balances": {
        "get": {
          "tags": [
            "2 Счета и балансы"
          ],
          "summary": "3. Получить баланс счета",
          "description": "Получение баланса счета\n\n**Требует:** Client token (для своих счетов) или Bank token с согласием (межбанк)",
          "operationId": "get_balances_accounts__account_id__balances_get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "account_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "description": "ID счета",
                "title": "Account Id"
              },
              "description": "ID счета",
              "example": "acc-1010"
            },
            {
              "name": "x-consent-id",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID согласия (получите через POST /account-consents/request). Обязателен для межбанковых запросов",
                "title": "X-Consent-Id"
              },
              "description": "ID согласия (получите через POST /account-consents/request). Обязателен для межбанковых запросов",
              "example": "consent-69e75facabba"
            },
            {
              "name": "x-requesting-bank",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID вашей команды (от организаторов). Укажите для запроса данных из другого банка",
                "title": "X-Requesting-Bank"
              },
              "description": "ID вашей команды (от организаторов). Укажите для запроса данных из другого банка",
              "example": "team200"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/accounts/{account_id}/transactions": {
        "get": {
          "tags": [
            "2 Счета и балансы"
          ],
          "summary": "4. Получить историю транзакций",
          "description": "Получение списка транзакций по счету\n\n**Пагинация:**\n- `page` — номер страницы (по умолчанию: 1)\n- `limit` — количество транзакций на странице (по умолчанию: 50, макс: 500)\n\n**Примеры:**\n- `GET /accounts/acc-1/transactions` — первые 50 транзакций\n- `GET /accounts/acc-1/transactions?page=2&limit=100` — следующие 100 транзакций\n- `GET /accounts/acc-1/transactions?limit=200` — первые 200 транзакций",
          "operationId": "get_transactions_accounts__account_id__transactions_get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "account_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "description": "ID счета",
                "title": "Account Id"
              },
              "description": "ID счета",
              "example": "acc-1010"
            },
            {
              "name": "from_booking_date_time",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "From Booking Date Time"
              },
              "example": "2025-01-01T00:00:00Z"
            },
            {
              "name": "to_booking_date_time",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "To Booking Date Time"
              },
              "example": "2025-12-31T23:59:59Z"
            },
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 1,
                "title": "Page"
              },
              "example": 1
            },
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "maximum": 100,
                "minimum": 1,
                "default": 50,
                "title": "Limit"
              },
              "example": 50
            },
            {
              "name": "x-consent-id",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID согласия (получите через POST /account-consents/request). Обязателен для межбанковых запросов",
                "title": "X-Consent-Id"
              },
              "description": "ID согласия (получите через POST /account-consents/request). Обязателен для межбанковых запросов",
              "example": "consent-69e75facabba"
            },
            {
              "name": "x-requesting-bank",
              "in": "header",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "ID вашей команды (от организаторов). Укажите для запроса данных из другого банка",
                "title": "X-Requesting-Bank"
              },
              "description": "ID вашей команды (от организаторов). Укажите для запроса данных из другого банка",
              "example": "team200"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "CreateAccountRequest": {
          "properties": {
            "account_type": {
              "type": "string",
              "title": "Account Type"
            },
            "initial_balance": {
              "type": "number",
              "title": "Initial Balance",
              "default": 0
            }
          },
          "type": "object",
          "required": [
            "account_type"
          ],
          "title": "CreateAccountRequest",
          "description": "Запрос на создание нового счета"
        },
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "HTTPBearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
...
  }